refactor: separate hooks & provider conponents
