+++ src/app/(auth)/signup/page.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect } from "react";
 import { useRouter } from "next/navigation";
-import { useAuth } from "@/context/AuthContext";
+import { useAuth } from "@/hooks/useAuth";
 import AuthForm from "@/components/AuthForm";
 
 export default function SignUpPage() {
