+++ src/components/Header.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { signOut } from "firebase/auth";
 import { auth } from "@/lib/firebase/config";
-import { useAuth } from "@/context/AuthContext"; // custom hook
+import { useAuth } from "@/hooks/useAuth"; // custom hook
 
 export default function Header() {
   const { user, loading } = useAuth();
