+++ src/components/providers/AuthProvider.tsx
@@ -1,0 +1,23 @@
+"use client";
+
+import { useEffect, useState, ReactNode } from "react";
+import { onAuthStateChanged, User } from "firebase/auth";
+import { auth } from "@/lib/firebase/config";
+import { AuthContext } from "@/context/AuthContext";
+
+export const AuthProvider = ({ children }: { children: ReactNode }) => {
+  const [user, setUser] = useState<User | null>(null);
+  const [loading, setLoading] = useState(true);
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, (user) => {
+      setUser(user);
+      setLoading(false);
+    });
+    return () => unsubscribe();
+  }, []);
+
+  const value = { user, loading };
+
+  return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
+};
