+++ src/context/AuthContext.tsx
@@ -1,40 +1,12 @@
-"use client";
+import { createContext } from "react";
+import type { User } from "firebase/auth";
 
-import { createContext, useContext, useEffect, useState, ReactNode } from "react";
-import { onAuthStateChanged, User } from "firebase/auth";
-import { auth } from "@/lib/firebase/config";
-
-type AuthContextType = {
+export type AuthContextType = {
   user: User | null;
   loading: boolean;
 };
 
-// initial context
-const AuthContext = createContext<AuthContextType>({
+export const AuthContext = createContext<AuthContextType>({
   user: null,
   loading: true,
 });
-
-export const AuthProvider = ({ children }: { children: ReactNode }) => {
-  const [user, setUser] = useState<User | null>(null);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    // onAuthStateChanged is automatic firebase listener
-    const unsubscribe = onAuthStateChanged(auth, (user) => {
-      setUser(user);
-      setLoading(false);
-    });
-
-    return () => unsubscribe();
-  }, []);
-
-  const value = { user, loading };
-
-  return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
-};
-
-// custom hook for context access
-export const useAuth = () => {
-  return useContext(AuthContext);
-};
