+++ src/hooks/useAuth.ts
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import { AuthContext } from "@/context/AuthContext";
+
+export const useAuth = () => {
+  return useContext(AuthContext);
+};
